raspberrypi_central/webapp/app/requirements.txt
django==3.0.7
psycopg2-binary==2.8.5
gunicorn==20.0.4
djangorestframework==3.11.0
django-cors-headers==3.4.0
django-debug-toolbar==2.2
factory-boy==2.12.0
faker==4.1.1
jinja2==2.11.2
paho-mqtt==1.5.0
celery==4.4.*
django-celery-beat==2.0.0
python-telegram-bot==12.8
pytz==2020.1
requests==2.24.*
django_extensions==3.0.9
Pillow==8.0.*
